requirements.txt
pandas==2.2.2
openpyxl==3.1.2
reportlab==4.1.0
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
google-cloud-storage==2.16.0
pydantic==2.7.4
python-dotenv==1.0.0
